Support duplicate plugin instances in test_list_from_entry_points
Account for the possibility of the isort plugin occurring more than
once in the plugin list, in test_list_from_entry_points.  This is needed
to avoid test failures on Gentoo where tests are run in a venv
with --system-site-packages.  As a result, if ini2toml is installed
already, it grabs all plugins twice — once from the venv, and then again
from system site-packages directory.